initial readabilty clean of api
